Move typeOf utils into separate file
