+++ src/utilities/EventEmitter.ts
@@ -1,3 +1,5 @@
+import { isArray, isFunction, isObject, isString } from './TypeOf.js';
+
 function callListener(eventThis: { event: string }, listener: Function, args: unknown[]) {
   try {
     listener.apply(eventThis, args);
@@ -53,27 +55,6 @@ export function removeListener(
 // Equivalent of Platform.config.inspect from ably-js for browser/RN
 function inspect(args: any): string {
   return JSON.stringify(args);
-}
-
-function typeOf(arg: unknown): string {
-  return Object.prototype.toString.call(arg).slice(8, -1);
-}
-
-// Equivalent of Util.isObject from ably-js
-function isObject(arg: unknown): arg is Record<string, unknown> {
-  return typeOf(arg) === 'Object';
-}
-
-function isFunction(arg: unknown): arg is Function {
-  return typeOf(arg) === 'Function';
-}
-
-function isString(arg: unknown): arg is String {
-  return typeOf(arg) === 'String';
-}
-
-function isArray<T>(arg: unknown): arg is Array<T> {
-  return Array.isArray(arg);
 }
 
 type EventMap = Record<string, any>;
